+++ Budget/app/src/main/java/com/example/budget/Summary.java
@@ -1,8 +1,8 @@
 package com.example.budget;
 
 public class Summary {
-    public float fTotalAsset;
-    public float fTotalLiability;
-    public float fTotalIncome;
-    public float fTotalExpense;
+    public Double fTotalAsset;
+    public Double fTotalLiability;
+    public Double fTotalIncome;
+    public Double fTotalExpense;
 }
